Proposal for a metrics/git.py file
This is a proposal for how to organize the classes and functions in a metrics file.

The functions at the end can later be organized in an a class, using the registry pattern, for example. But for now, we can have just functions. The report would call those functions, with the corresponding index as argument, and would receive in return the values for the metrics to be computed.

This is just a draft, I didn't intend to run it. But maybe you can get the idea...